Refactor time manager code a bit, to allow alternative time managers. (#1163)
* WIP

* Some more changes.

* Don't ignore move-overhead parameter for `go movetime` searches.

* Added a comment.
