+++ meson.build
@@ -122,7 +122,9 @@ files += [
   'src/mcts/node.cc',
   'src/mcts/params.cc',
   'src/mcts/search.cc',
+  'src/mcts/stoppers/common.cc',
   'src/mcts/stoppers/factory.cc',
+  'src/mcts/stoppers/legacy.cc',
   'src/mcts/stoppers/stoppers.cc',
   'src/mcts/stoppers/timemgr.cc',
   'src/neural/cache.cc',
