+++ src/mcts/stoppers/legacy.h
@@ -1,6 +1,6 @@
 /*
   This file is part of Leela Chess Zero.
-  Copyright (C) 2019-2020 The LCZero Authors
+  Copyright (C) 2020 The LCZero Authors
 
   Leela Chess is free software: you can redistribute it and/or modify
   it under the terms of the GNU General Public License as published by
@@ -27,17 +27,11 @@
 
 #pragma once
 
-#include "mcts/stoppers/common.h"
-#include "mcts/stoppers/timemgr.h"
 #include "utils/optionsdict.h"
-#include "utils/optionsparser.h"
 
 namespace lczero {
 
-// Populates UCI/command line flags with time management options.
-void PopulateTimeManagementOptions(RunType for_what, OptionsParser* options);
-
-// Creates a new time manager for a new search.
-std::unique_ptr<TimeManager> MakeTimeManager(const OptionsDict& dict);
+std::unique_ptr<TimeManager> MakeLegacyTimeManager(int64_t move_overhead,
+                                                   const OptionsDict& params);
 
 }  // namespace lczero
